add timer to more methods
